Changes to Device Streaming Proxy sample to allow for reconnections
NOTE: These changes are not intendend to be merged back into the
DeviceStreaming branch!!

This is just a demo of how to change the proxy sample to allow
successive unique connections to the sample sample runtime instance,
making it reusable (i.e., the sample will not terminate when the first
streaming pipe connection is ended).

These changes still allow only one connection to be in progress at any
given time. Attemping to connect again to the "proxy" will kill an existing streaming connection.